Scan/requirements.txt
python-nmap
tqdm
pyfiglet
termcolor

